+++ app/src/main/java/com/example/themoviedb/retrofitservice/requests/models/MoviesModel.kt
@@ -37,7 +37,7 @@ class MoviesModel {
     @SerializedName("release_date")
     val releaseDate: String? = null
     @SerializedName("revenue")
-    val revenue: Int? = null
+    val revenue: Long? = null
     @SerializedName("runtime")
     val runtime: Int? = null
     @SerializedName("spoken_languages")
